[IMP] partner_serializer: New addon providing an absract model in charge of managing the serialization of a partner to an address
