+++ shopinvader_sale_cart/__init__.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
 from . import services
+from . import models
