+++ shopinvader_sale_cart/models/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+from . import res_partner_serializer
